Cron jobs are not removed when unused.
Currently, if you have the bad idea of setting backup_cron or
optimize_cron to true in the main class and later to change it back to
false, the associated cronjob will stay in place even though you don't
want it.

To make this option more consistant, let's remove cron jobs (and backup
helper scripts) when the appropriate option is set to false.

Signed-off-by: Gabriel Filion <gabster@lelutin.ca>